refactor : rename MN, FD, LV, FM
