--- Assets/Scenes/FD.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 89b39684595c0d44893d5008af91cded
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
