--- Assets/Scenes/FM.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b539453e4c628f94d9f9bc8280397a6d
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
