+++ Assets/Scenes/LoverScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8a9b9cd94b8b8aa4bb6cf05bd0a9545f
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
