--- Assets/Scenes/MN.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: c76d4ac4fbfb5814a84e0e86197b5010
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
