+++ Assets/Scenes/StreetScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c76d4ac4fbfb5814a84e0e86197b5010
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
