--- Assets/Scripts/Game/FD/FD_Scene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b762f891466e3c4fbc0241b1f460dbf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
