+++ Assets/Scripts/Game/FD/FriendSceneFlow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 433947a72a597ca49856b59dc827b5a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
