--- Assets/Scripts/Game/FM/FM_GameFramework.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b9017b12946a1b24797e4db689060619
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
