--- Assets/Scripts/Game/FM/FM_Scene.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7a1e6808b1023164d972610eca262ec0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
