--- Assets/Scripts/Game/LV/LV_GameFramework.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2b177ddc856a3254587dd5ab61a34f7d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
