+++ Assets/Scripts/Game/LV/LoverScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d841b6031be6dba4ab0ef32c8fe2b9b2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
