+++ Assets/Scripts/Game/LV/LoverSceneFlow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0d2e1068066642c408597e57522185b7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
