--- Assets/Scripts/Game/StartRoom.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8e82ccd2b45d2c1428a51aa47d9bcc5b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
