--- Assets/Scripts/Game/StartRoom/StartRoomEvent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0b2c18b505b45ab45ae8f407a5ac662e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
