--- Assets/Scripts/Game/StartRoom/StartRoomGameFramework.cs
@@ -1,10 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using DG.Tweening;
-using UnityEngine;
-
-public class StartRoomGameFramework : MonoBehaviour
-{
-    
-}
