--- Assets/Scripts/Game/StartRoom/StartRoomGameFramework.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 75cfa7ecaa5e0ac41a5c8ee38c90800c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
