+++ Assets/Scripts/Game/Street.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e82ccd2b45d2c1428a51aa47d9bcc5b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
