+++ Assets/Scripts/Game/Street/StreetScene.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c5190a0000b1b2e4e89e9f9761cd63e1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
