+++ Assets/Scripts/Game/Street/StreetSceneFlow.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using DG.Tweening;
+using UnityEngine;
+
+public class StreetSceneFlow : MonoBehaviour
+{
+    
+}
