+++ Assets/Scripts/Game/Street/StreetSceneFlow.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fa200c340c9a89b44bd81772b9c62a38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
