+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,16 +6,16 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/MN.unity
+    path: Assets/Scenes/StreetScene.unity
     guid: c76d4ac4fbfb5814a84e0e86197b5010
   - enabled: 1
-    path: Assets/Scenes/FD.unity
+    path: Assets/Scenes/FriendScene.unity
     guid: 89b39684595c0d44893d5008af91cded
   - enabled: 1
-    path: Assets/Scenes/LV.unity
+    path: Assets/Scenes/LoverScene.unity
     guid: 8a9b9cd94b8b8aa4bb6cf05bd0a9545f
   - enabled: 1
-    path: Assets/Scenes/FM.unity
+    path: Assets/Scenes/FamilyScene.unity
     guid: b539453e4c628f94d9f9bc8280397a6d
   - enabled: 1
     path: Assets/DebugOnly/SampleScene.unity
